Make k9mail module specify which Backend implementations to include
