+++ app/core/src/main/java/com/fsck/k9/Core.kt
@@ -9,7 +9,6 @@ import android.os.Handler
 import android.os.Looper
 import android.os.StrictMode
 import com.fsck.k9.autocrypt.autocryptModule
-import com.fsck.k9.backend.backendModule
 import com.fsck.k9.crypto.openPgpModule
 import com.fsck.k9.mail.internet.BinaryTempFileBody
 import com.fsck.k9.mail.ssl.LocalKeyStore
@@ -38,7 +37,6 @@ object Core : KoinComponent {
             openPgpModule,
             autocryptModule,
             mailStoreModule,
-            backendModule,
             extractorModule,
             htmlModule,
             coreNotificationModule
