+++ app/k9mail/src/main/java/com/fsck/k9/backends/ImapBackendFactory.kt
@@ -1,32 +1,45 @@
-package com.fsck.k9.backend
+package com.fsck.k9.backends
 
 import android.content.Context
+import android.net.ConnectivityManager
 import com.fsck.k9.Account
 import com.fsck.k9.Preferences
+import com.fsck.k9.backend.BackendFactory
 import com.fsck.k9.backend.api.Backend
-import com.fsck.k9.backend.pop3.Pop3Backend
+import com.fsck.k9.backend.imap.ImapBackend
 import com.fsck.k9.mail.ServerSettings
 import com.fsck.k9.mail.oauth.OAuth2TokenProvider
+import com.fsck.k9.mail.power.PowerManager
 import com.fsck.k9.mail.ssl.DefaultTrustedSocketFactory
-import com.fsck.k9.mail.store.pop3.Pop3Store
+import com.fsck.k9.mail.store.imap.ImapStore
 import com.fsck.k9.mail.transport.smtp.SmtpTransport
 import com.fsck.k9.mail.transport.smtp.SmtpTransportUriCreator
 import com.fsck.k9.mail.transport.smtp.SmtpTransportUriDecoder
 import com.fsck.k9.mailstore.K9BackendStorage
 
-class Pop3BackendFactory(private val context: Context, private val preferences: Preferences) : BackendFactory {
+class ImapBackendFactory(
+        private val context: Context,
+        private val preferences: Preferences,
+        private val powerManager: PowerManager
+) : BackendFactory {
     override val transportUriPrefix = "smtp"
 
     override fun createBackend(account: Account): Backend {
         val accountName = account.description
         val backendStorage = K9BackendStorage(preferences, account, account.localStore)
-        val pop3Store = createPop3Store(account)
+        val imapStore = createImapStore(account)
         val smtpTransport = createSmtpTransport(account)
-        return Pop3Backend(accountName, backendStorage, pop3Store, smtpTransport)
+        return ImapBackend(accountName, backendStorage, imapStore, powerManager, smtpTransport)
     }
 
-    private fun createPop3Store(account: Account): Pop3Store {
-        return Pop3Store(account, DefaultTrustedSocketFactory(context))
+    private fun createImapStore(account: Account): ImapStore {
+        val oAuth2TokenProvider: OAuth2TokenProvider? = null
+        return ImapStore(
+                account,
+                DefaultTrustedSocketFactory(context),
+                context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager,
+                oAuth2TokenProvider
+        )
     }
 
     private fun createSmtpTransport(account: Account): SmtpTransport {
@@ -35,11 +48,11 @@ class Pop3BackendFactory(private val context: Context, private val preferences: 
     }
 
     override fun decodeStoreUri(storeUri: String): ServerSettings {
-        return Pop3Store.decodeUri(storeUri)
+        return ImapStore.decodeUri(storeUri)
     }
 
     override fun createStoreUri(serverSettings: ServerSettings): String {
-        return Pop3Store.createUri(serverSettings)
+        return ImapStore.createUri(serverSettings)
     }
 
     override fun decodeTransportUri(transportUri: String): ServerSettings {
