+++ app/k9mail/src/main/java/com/fsck/k9/backends/Pop3BackendFactory.kt
@@ -1,44 +1,33 @@
-package com.fsck.k9.backend
+package com.fsck.k9.backends
 
 import android.content.Context
-import android.net.ConnectivityManager
 import com.fsck.k9.Account
 import com.fsck.k9.Preferences
+import com.fsck.k9.backend.BackendFactory
 import com.fsck.k9.backend.api.Backend
-import com.fsck.k9.backend.imap.ImapBackend
+import com.fsck.k9.backend.pop3.Pop3Backend
 import com.fsck.k9.mail.ServerSettings
 import com.fsck.k9.mail.oauth.OAuth2TokenProvider
-import com.fsck.k9.mail.power.PowerManager
 import com.fsck.k9.mail.ssl.DefaultTrustedSocketFactory
-import com.fsck.k9.mail.store.imap.ImapStore
+import com.fsck.k9.mail.store.pop3.Pop3Store
 import com.fsck.k9.mail.transport.smtp.SmtpTransport
 import com.fsck.k9.mail.transport.smtp.SmtpTransportUriCreator
 import com.fsck.k9.mail.transport.smtp.SmtpTransportUriDecoder
 import com.fsck.k9.mailstore.K9BackendStorage
 
-class ImapBackendFactory(
-        private val context: Context,
-        private val preferences: Preferences,
-        private val powerManager: PowerManager
-) : BackendFactory {
+class Pop3BackendFactory(private val context: Context, private val preferences: Preferences) : BackendFactory {
     override val transportUriPrefix = "smtp"
 
     override fun createBackend(account: Account): Backend {
         val accountName = account.description
         val backendStorage = K9BackendStorage(preferences, account, account.localStore)
-        val imapStore = createImapStore(account)
+        val pop3Store = createPop3Store(account)
         val smtpTransport = createSmtpTransport(account)
-        return ImapBackend(accountName, backendStorage, imapStore, powerManager, smtpTransport)
+        return Pop3Backend(accountName, backendStorage, pop3Store, smtpTransport)
     }
 
-    private fun createImapStore(account: Account): ImapStore {
-        val oAuth2TokenProvider: OAuth2TokenProvider? = null
-        return ImapStore(
-                account,
-                DefaultTrustedSocketFactory(context),
-                context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager,
-                oAuth2TokenProvider
-        )
+    private fun createPop3Store(account: Account): Pop3Store {
+        return Pop3Store(account, DefaultTrustedSocketFactory(context))
     }
 
     private fun createSmtpTransport(account: Account): SmtpTransport {
@@ -47,11 +36,11 @@ class ImapBackendFactory(
     }
 
     override fun decodeStoreUri(storeUri: String): ServerSettings {
-        return ImapStore.decodeUri(storeUri)
+        return Pop3Store.decodeUri(storeUri)
     }
 
     override fun createStoreUri(serverSettings: ServerSettings): String {
-        return ImapStore.createUri(serverSettings)
+        return Pop3Store.createUri(serverSettings)
     }
 
     override fun decodeTransportUri(transportUri: String): ServerSettings {
