+++ backend/pop3/build.gradle
@@ -12,8 +12,8 @@ dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:${versions.kotlin}"
 
     api project(":backend:api")
-    implementation project(":mail:protocols:pop3")
-    implementation project(":mail:protocols:smtp")
+    api project(":mail:protocols:pop3")
+    api project(":mail:protocols:smtp")
 
     implementation "com.jakewharton.timber:timber:${versions.timber}"
